Add slow-fixture seam + progress.live marker for live-progress UI tests
DEBUG-only per-file delay (--testPerFileDelayMs, gated on UITestSeam.isActive)
keeps the copy and verify phases on screen; a progress.live accessibility
marker on the Backup Progress headline exposes phase/overall/per-dest counts to
the XCUITest suite. Split the four progress detail views out of
MultiDestinationProgressSection.swift to stay under the 500-line limit.

Implements AMUX-461.
